style(users): optimize interface business logic into milldeware
